feat: added a test case to cover custom function in backend-auth
